[Test] API 응답통일 테스트 코드 구현
