--- src/main/java/F0/Suzip/controller/TestController.java
@@ -1,13 +1,0 @@
-package F0.Suzip.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TestController {
-
-    @GetMapping("/test")
-    public String test() {
-        return "success";
-    }
-}
