+++ src/main/java/Fo/Suzip/converter/TempConverter.java
@@ -1,0 +1,11 @@
+package F0.Suzip.converter;
+
+import F0.Suzip.web.dto.TempResponse;
+
+public class TempConverter {
+    public static TempResponse.TempTestDTO toTempTestDTO(){
+        return TempResponse.TempTestDTO.builder()
+                .testString("This is Test!")
+                .build();
+    }
+}
