+++ src/main/java/Fo/Suzip/web/dto/TempResponse.java
@@ -1,0 +1,16 @@
+package F0.Suzip.web.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+public class TempResponse {
+    @Builder
+    @Getter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class TempTestDTO{
+        String testString;
+    }
+}
